chore: consolidate navigation controller code
